Refactor and enhance firmware functionality
- Updated DeviceShadowStoreTest to accommodate additional parameters in DeviceState.
- Modified MqttMessageHandlingIntegrationTest to reflect changes in sensor data structure and added assertions for pump and light relay states.
- Changed HardwareProfile to enable additional features.
- Initialized RTC provider in AppRuntime and linked it with TimeService.
- Updated Event structure to use uint64_t for value to support larger timestamps.
- Implemented DS3231 RTC driver with comprehensive time reading functionality.
- Enhanced ActuatorModule to utilize TimeService for accurate event timestamps.
- Adjusted AutomationModule to handle event timestamps correctly and improved payload formatting.
- Updated SensorHubModule to ensure proper state serialization and added tests for pump status and timestamps.
- Enhanced TimeService with methods for better time validation and retrieval.
- Added Ds3231RtcProvider to interface with the DS3231 RTC chip.
- Introduced unit tests for new functionalities, ensuring robust validation of state and time handling.

+++ firmware/src2/core/EventQueue.h
@@ -44,7 +44,7 @@ struct Event {
   // Tip sobytiya.
   EventType type;
   // Dop. chislennoe znachenie sobytiya.
-  uint32_t value;
+  uint64_t value;
   // MQTT dannye dlya sobytiy tipa kMqttMessage.
   MqttMessage mqtt;
 };
